requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database Drivers
psycopg2-binary>=2.9.9
neo4j>=5.15.0
sqlalchemy>=2.0.0
geoalchemy2>=0.14.0

# Data Processing
pandas>=2.1.0
geopandas>=0.14.0
numpy>=1.24.0

# API and Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0

# NLP and LLM (for Text2Cypher)
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.2

# Visualization and Reporting
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code Quality
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

# Utilities
loguru>=0.7.2
tqdm>=4.66.0
click>=8.1.7
rich>=13.7.0

# Development
jupyter>=1.0.0
ipython>=8.17.0
ipykernel>=6.26.0
